+++ DeepSkyStacker/StackingTasks.cpp
@@ -869,7 +869,7 @@ public :
 private :
 	void	AdjustValue(double & fValue)
 	{
-		fValue = min(max(0, fValue), 255.0);
+		fValue = min(max(0.0, fValue), 255.0);
 	};
 
 public :
@@ -1890,16 +1890,16 @@ __int64	CAllStackingTasks::ComputeNecessaryDiskSpace(CRect & rcOutput)
 				m_vStacks[i].m_pLightTask->m_Method = MBP_AVERAGE;
 
 			if (m_vStacks[i].m_pOffsetTask)
-				ulOffsetSpace = max(ulOffsetSpace, ulSpace * m_vStacks[i].m_pOffsetTask->m_vBitmaps.size());
+				ulOffsetSpace = max(ulOffsetSpace, static_cast<__int64>(ulSpace * m_vStacks[i].m_pOffsetTask->m_vBitmaps.size()));
 
 			if (m_vStacks[i].m_pDarkTask)
-				ulDarkSpace = max(ulDarkSpace, ulSpace * m_vStacks[i].m_pDarkTask->m_vBitmaps.size());	
+				ulDarkSpace = max(ulDarkSpace, static_cast<__int64>(ulSpace * m_vStacks[i].m_pDarkTask->m_vBitmaps.size()));
 
 			if (m_vStacks[i].m_pDarkFlatTask)
-				ulDarkFlatSpace = max(ulDarkFlatSpace, ulSpace * m_vStacks[i].m_pDarkFlatTask->m_vBitmaps.size());	
+				ulDarkFlatSpace = max(ulDarkFlatSpace, static_cast<__int64>(ulSpace * m_vStacks[i].m_pDarkFlatTask->m_vBitmaps.size()));
 
 			if (m_vStacks[i].m_pFlatTask)
-				ulFlatSpace = max(ulFlatSpace, ulSpace * m_vStacks[i].m_pFlatTask->m_vBitmaps.size());	
+				ulFlatSpace = max(ulFlatSpace, static_cast<__int64>(ulSpace * m_vStacks[i].m_pFlatTask->m_vBitmaps.size()));
 		};
 	};
